Raw quoting syntax for first-class subshells
Other shells allow creating subshells using `(cd somewhere && do-something)`.
Our answer to this is to create a subshell like a regular process (#1439):

    $SHELL -c 'cd somewhere && do-something'

This syntax works the same way for non-shell programs

    watch 'ps aux | grep foo'
    git bisect run $SHELL -c 'make || exit 125; make test'
    hyperfine '...'

Issues with our solution are:
- the argument has no command-highlighting, completions (#9406) or indentation
- the  user may need to escape the argument multiple times ("quoting hell").
  This primarily affects arguments containing backslashes and quotes
  (for example «$SHELL -c 'printf %s\\\n *'»).
  In many other cases backslash escaping is not necessary however the rules
  can be confusing.

One workaround is to compose the subcommand in a separate shell prompt and
copy-paste it (inside quotes for automatic escaping).
That's not ideal because it breaks the writer's flow, and it doesn't work
well if the command needs to be refined.

Add a `(( ))` raw quoting syntax for subshells and deferred commands.
Syntax highlighting, completion and indentation work the same as in top-level
commands. No expansion/unescaping takes place:

    $SHELL -c ((printf %s\\n *))

Nested parentheses are okay, similar to Tcl's braced strings.
Imbalanced quotation marks are an error.
This means that the quoted string cannot contain an imbalance of its quote
characters.
(The good side is that, in many cases, imbalance is not intentional).

If we stick to this syntax, it would be good to also support `(< >)` and
`({ })` with the same meaning as `(( ))`, to make it easier to use when the
quoted content contains an unmatched `(`.

---

Also add a `{{ }}` raw quoting syntax that does not get shell highlighting,
completion and indentation.
This is meant for languages like Perl, awk, Python and regex dialects (no
more need to escape backslashes).
The would be a breaking change, we might want to avoid that by picking a
different syntax.

To-do:
- figure out how we want to tweak syntax, possibly find something that can
  be added to Bash.
- adjust __fish_tokenizer_state
- docs/changelog/tests