Use ahash for our own HashMap and HashSet objects
We already transitively take a dependency on the ahash crate, so might as well
use it. While not all of our HashMap/HashSet objects are performance-sensitive,
some of them are and there's no cost to migrating them all over.

By default, rust uses a DoS-resistant, cryptographic hash (SipHash) to hash the
keys for HashMap/HashSet entries into buckets, which is slower than what we used
to use in the C++ world. We don't use any of it in a security-sensitive context
or based off of unvalidated/untrusted/remote input, so there are no security
reasons to continue to use SipHash.

Some of fish's HashMap/HashSet instances could benefit from being faster, e.g.
the functions cache, various completion caches, environment variable tables, and
history lookups, so this isn't just a "use it because it's there" kind of thing.