Add `set --keep` to avoid overwriting an existing variable
This is my first time contributing to fish, so please point out any flaws in the procedure.
Also, I would not consider myself a very power user of fish, so I might be missing something obvious, too.

This adds support for a `-k` or `--keep` option to the `set` command.
When used in a `set` invokation that would set a variable to a value, the option turns the command into a no-op if a variable is already there.

I was looking at the changes in 3.7.1 after noticing a nice prompt in a fossil checkout of mine.
This lead me to read the fossil prompt function and to stumble upon [this pattern](https://github.com/fish-shell/fish-shell/blob/7850142bef794db4d031fb6e1ec9dd7906a9f459/share/functions/fish_fossil_prompt.fish#L11-L47).

I thought there must be a better way... so here it is: `set -k -g foo bar` is (supposed to be) equivalent to `set -q foo or set -g foo bar`.