Add support for `string unescape --style=regex`
This complements existing support for `string escape --style=regex`.

PCRE2's spec is rather extensive and supports a massive number of escape
sequences, but we are only interested in those that are formed by escaping a
literal string (i.e. not actually regexes).

In addition to reversing the effects of `string escape --style=regex`, support
for unescaping some common "in-the-wild" escape variants is also added. For
example, `printf "hello\nworld\n"` is passed through unmodified by `string
escape` but it's easy to imagine that a different regex escape library would
encode that as the (fish-encoded) `hello\\nworld\\n` (escaping the new lines).
We support both, since the goal here isn't to strictly unmap fish's `string
escape` output (which would be of limited utility) but to also interact with
regex expressions found in files or emitted by executables in the course of
writing shell scripts (such as completions).

Closes #10611.